Using Ebean. Moved JPA to test.

+++ app/dto/PageDto.java
@@ -8,5 +8,5 @@ import java.util.List;
 @Data
 public class PageDto <T> {
     private final List<T> items;
-    private final Long totalCount;
+    private final int totalCount;
 }
