+++ app/models/repositories/Page.java
@@ -7,5 +7,5 @@ import java.util.List;
 @Data
 public class Page <T> {
     private final List<T> items;
-    private final long totalCount;
+    private final int totalCount;
 }
